Added reading functions and fixed issue #6
Issue # 6 was that createchar() function leave data address in CGRAM position. After that, begin(), setCursor(),clear() or home() functions must be executed before write() function in order for it to work. It is fixed but require rw pin to be set for it to work. This is because, when address register goes to CGRAM, there is no record of its previous position in our MCU. MCU must read its previous position in order for it to go back. In case rw pin is not specified, the register is not returned and it cannot untill rw is specified.
Also Added following functions,
read_BF_addr
read_RAM
read
read4bits
read8bits