Remove redundant spaces from comment text
Some redundant spaces were present in the comments

Some of these followed a period. Although two spaces after a period were sometimes recommended in previous eras, modern
practice is universally a single space:

https://www.chicagomanualofstyle.org/qanda/data/faq/topics/OneSpaceorTwo/faq0006.html

Regardless of whether it is correct according to standard practice, it might be reasonable to retain double spaces if
this were a well established convention in the project. That is not the case here. The majority of the content uses only
a single space, so the double spaces are inconsistent in addition to being incorrect.

There were also some occurrences of redundant spaces within sentences, which is never correct.