Mark character lines array as const
Characters may be defined as static constexpr arrays in a client application, but passing the data to the LiquidLibrary requires a const_cast because it is not flagged const, even if in practice uint8_t are not modified. This PR will allow to remove the const_cast in client code.